completion API endpoint for mocks take 5 seconds. Needs to create `/slots` endpoint.
